(GH-445) Chocolatey-Version improvements
This should help improve the functionality of chocolatey-version by setting it
to use the improved Chocolatey-List should speed some things up.

The perf when calling cver all is still going to be just as bad, there are
minor improvements but it makes a call to nuget for every package to check
for a newer version because nuget is limited in what you can pass to it. This
should have significant improvements when we change over to using pkg indexes
as they will be almost instantaneous returns.