Get-ChocolateyUnzip: fix problems accessing paths under System32 on 64-bit
On a 64-bit OS, as 7za.exe is a 32-bit application, when it is passed
paths pointing inside $Env:SystemRoot\System32 and uses those paths in
Windows API calls, the paths get transparently translated by the WOW64
filesystem redirector to $Env:SystemRoot\SysWOW64\...

This happens, for instance, when Chocolatey is executed as SYSTEM (e.g.
from a service). A real-life example is using Chocolatey to configure
Azure virtual machines using the startup script feature.

The SYSTEM identity has TEMP pointing to
$Env:SystemRoot\System32\config\systemprofile\AppData\Local\Temp.
That path is used by Chocolatey for storage of downloaded archive files
and is passed to 7za.exe when extracting those files. Because of WOW64
redirection, 7za.exe looks for the files in the wrong place and fails.

The solution is to replace 'System32' with 'sysnative' in paths passed to
32-bit processes. The sysnative directory, accessible only from 32-bit
processes, points to the 64-bit system directory.