(GH-153)(GH-134) Update PATH on cmd.exe
This adds a command RefreshEnv.cmd to the bin files, so that one can update
environment variables out of band. There is zero chance it will work once
chocolatey.cmd calls exit /b %ERRORLEVEL% as exiting the batch effectively
destroys any updates to local environmental changes. The way it currently
works is effectively the same as using setlocal due to the use of exit.

Since this command can be called out of band, it should be called out of
band, whenever the user wants to refresh their environment variables they
can call this. Note this only works in cmd.exe shell. There is no current way
to update a powershell session other than importing the helper functions as
a module and running Update-SessionEnvironment.