(GH-516) If process has not yet exited, wait the process
The prior attempts at waiting the process resulted in error when the process
had already exited. This caused really poor behavior and things to error when
they should have been ignored. Further wrapping wait process in try catch is
not enough as it will ignore the catch clause, so check to see if the process
has exited first prior to continuing execution.